+++ docs/releasenotes/releasenotes.rst
@@ -22,6 +22,8 @@ Enhancements
       filter: '1 month',   // to only show news items from last month
     }
 
+  * New special block: alarmmeldingen (Dutch). See <todo>
+
 Fixes
 
   * Fix for Evo devices